+++ CHANGELOG.rst
@@ -11,9 +11,10 @@ Added:
 - autoupdate functionality
 - switched installer to per-user location
 - removed machine_vars_path from nautilai.toml and hard-coded to C:\ProgramData\Curi Bio\Nautilai\machine.toml
+- Plate ID widget
 
 Changed:
-^^^^^^
+^^^^^^^^
 - Nautilai rename
 - Removal of default config values
 - Local Analysis now outputs time series data as parquet